Can print out center of markers

+++ brain/cam.hh
@@ -4,7 +4,7 @@
 #include "opencv2/highgui.hpp"
 #include "opencv2/imgproc.hpp"
 #include<vector>
-typedef std::pair<int, int> Posn; 
+typedef std::pair<float, float> Posn; 
 void cam_init();
 void cam_show(cv::Mat);
 void detect_markers(cv::Mat frame);
